refactor(model): 将博客字数字段类型从 uint16 改为 uint64
- 在 dto、po 和 vo 文件中，将 BlogWordsNum 字段的类型从 uint16 改为 uint64
- 这个改动可以支持更高字数的博客，避免潜在的整数溢出问题
- 删除了 adminrouter 中的错误处理代码

+++ internal/model/po/po.go
@@ -9,7 +9,7 @@ type Blog struct {
 	BlogBrief    string    `gorm:"column:blog_brief"`                                           // 博客简介
 	CategoryId   string    `gorm:"column:category_id"`                                          // 逻辑外键字段（无约束）
 	BlogState    bool      `gorm:"column:blog_state"`                                           // 博客状态
-	BlogWordsNum uint16    `gorm:"column:blog_words_num"`                                       // 博客字数
+	BlogWordsNum uint64    `gorm:"column:blog_words_num"`                                       // 博客字数
 	BlogIsTop    bool      `gorm:"column:blog_is_top"`                                          // 是否置顶
 	CreateTime   time.Time `gorm:"column:create_time;default:CURRENT_TIMESTAMP"`                // 创建时间
 	UpdateTime   time.Time `gorm:"column:update_time;default:CURRENT_TIMESTAMP;autoUpdateTime"` // 更新时间
