+++ internal/model/vo/vo.go
@@ -15,7 +15,7 @@ type BlogVo struct {
 	Category     *CategoryVo `json:"category,omitempty"`
 	Tags         []TagVo     `json:"tags,omitempty"`
 	BlogState    bool        `json:"blog_state"`
-	BlogWordsNum uint16      `json:"blog_words_num,omitempty"`
+	BlogWordsNum uint64      `json:"blog_words_num,omitempty"`
 	BlogIsTop    bool        `json:"blog_is_top"`
 	CreateTime   time.Time   `json:"create_time,omitempty"`
 	UpdateTime   time.Time   `json:"update_time,omitempty"`
